# Mark mhlo.broadcast op as indexable to clone into its consumers.